==> notebook/01_data_fetch.ipynb <==
Import statements

from pathlib import Path
import pandas as pd
import requests

def fetch_raw_data(year: int, month: int) -> str: #month:02 to specify month format as 01, 02, etc in the next line
    url = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02}.parquet" 
    response = requests.get(url)
    if response.status_code == 200: #Response status = OK Meaning the request was successful
        path = Path("..") / "data" / "raw" / f"rides_{year}_{month:02}.parquet" #Builds Path
        path.parent.mkdir(parents=True, exist_ok=True) #Makes sure the directory exists before savign file
        path.write_bytes(response.content)
        print(f"Successfully fetched: {str(path)}")
        return str(path)
    else:
        raise Exception(f"{url} is not available")


Function Test

fetch_raw_data(2023, 1) #Function test
