==> notebook/02_validate_and_save.ipynb <==
Import Statements

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"
#Displays all the results, not just the last one

Dataframe Conversion

import pandas as pd
from pathlib import Path
import pyarrow.parquet as pq

month = 1
year = 2023
path = Path("..") / "data" / "raw" / f"rides_{year}_{month:02}.parquet"

table = pq.read_table(path)
rides = table.to_pandas()
rides.head()
rides.shape
#Converst the downloaded data in the data/raw folder to pd dataframe

Duration and Total Amount Filter

rides_cp = rides.copy()
rides_cp["duration"] = rides["tpep_dropoff_datetime"] - rides["tpep_pickup_datetime"]
rides_cp.head()
rides_cp.shape
#Finds the duration using fropoff time and pickup time

duration_filter = (rides_cp["duration"] > pd.Timedelta(0)) & (rides_cp["duration"] <= pd.Timedelta(hours=5)) #Finds rows where duration < 0 or more than 5 hours and assigns True between 0 and 5
invalid_duration = sum(~duration_filter) / rides_cp.shape[0] * 100 #Checks %of invalid rows present
print(f"Invalid rows: {invalid_duration:.2f}%") #Prints % of invalid rows from duration

total_amount_filter = (rides_cp["total_amount"]  > 0) & (rides_cp["total_amount"] <= rides_cp["total_amount"].quantile(0.999))
invalid_total_amount = sum(~total_amount_filter) / rides_cp.shape[0] * 100 #Checks %of invalid rows present
print(f"Invalid rows: {invalid_total_amount:.2f}%") #Prints % of invalid rows from amount

nyc_locations = ~rides_cp["PULocationID"].isin((1, 264, 265))
invalid_locations = sum(~nyc_locations) / rides_cp.shape[0] * 100
print(f"Invalid rows: {invalid_locations:.2f}%") #Prints % of invalid rows from locations

filter_date_range = (rides_cp["tpep_pickup_datetime"] >= "2023-01-01") & (rides_cp["tpep_pickup_datetime"] < "2023-02-01")
invalid_date_range = sum(~filter_date_range) / rides_cp.shape[0] * 100
print(f"Invalid rows: {invalid_date_range:.2f}%") #Prints % of invalid rows from date_range


final_filter = duration_filter & total_amount_filter & nyc_locations & filter_date_range
numbers_dropped = final_filter.shape[0] - sum(final_filter) # numbers dropped
numbers_dropped
rows_dropped = numbers_dropped/final_filter.shape[0] * 100
print(f"Invalid rows: {rows_dropped:.2f}%") #Prints % of rows dropped using all filters

rides = rides[final_filter] #Apply final filter to rides
rides = rides[["tpep_pickup_datetime", "PULocationID"]]
rides.rename(columns={
    "tpep_pickup_datetime": "pickup_datetime",
    "PULocationID": "pickup_location_id"
}, inplace=True) #Filter and rename only two columns
rides.head()
year = 2023
month = 1
path = Path("..") / "data" / "processed" / f"rides_{year}_{month:02}.parquet"
rides.to_parquet(path, engine="pyarrow", index=False) #Save Processed parquet file
