==> notebook/05_transform__raw_into_features_targets.ipynb <==
Import statements and Autoreload

%load_ext autoreload
%autoreload 2
#All modules are autoreloaded
#(2 specifies all, even imported ones) when any change is made

import sys
import os
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))
from src.data_utils import load_and_process_taxi_data
from src.data_utils import transform_raw_data_into_ts_data
from src.data_utils import transform_ts_data_info_features_and_target_loop

from src.config import TRANSFORMED_DATA_DIR


Loading data for 2023, transforming, and adding attributes and target

rides = load_and_process_taxi_data(year=2023)
rides.head(5)
rides.shape
#Loads taxi data 2023

ts_data = transform_raw_data_into_ts_data(rides)
ts_data.head(5)
ts_data.shape
#Transforms loaded data (truncate hour, aggregate, fill missing entries)

features, targets = transform_ts_data_info_features_and_target_loop(ts_data, window_size=24*28*1, step_size=24)
features.head(5)
features.shape
#Adds features and target using window lead

Merging the features and target (along with shape check)

tabular_data = features.copy() 
#Using copy is needed here as otherwise features also gets updated
tabular_data["target"] = targets
features.shape
targets.shape
tabular_data.shape


Saving the merged df tabular data to transformed dir in parquet format

tabular_data.to_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet", engine="pyarrow")
