==> notebook/06_visualisations.ipynb <==
Import Statements

%load_ext autoreload
%autoreload 2

import sys
import os
import pandas as pd
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

from src.config import TRANSFORMED_DATA_DIR
from src.plot_utils import plot_aggregated_time_series

Loading merged parquet file and seperating the attributes and target

features_and_targets = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
features = features_and_targets.iloc[:, :-1]
targets = features_and_targets["target"]
features.head(5)
features.shape
targets.head(5)
targets.shape
#targets_df = pd.DataFrame(targets)
#targets_df["pickup_location_id"] = features["pickup_location_id"]
#targets_df.head(5)
#targets_df.shape
features[features['pickup_location_id']==43]

plot_aggregated_time_series(features, targets, row_id=3772)
