==> notebook/12_load_features_hopswork.ipynb <==
Import Statements

from dotenv import load_dotenv
import os
load_dotenv()
import sys
from datetime import datetime
import pandas as pd
import hopsworks

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

from src.data_utils import load_and_process_taxi_data
from src.data_utils import transform_raw_data_into_ts_data


Loading data using load_and_process function and transforming

from_year = 2023
# to_year = datetime.now().year
to_year = 2024
print(f"Download raw data from {from_year} to {to_year}")

rides = pd.DataFrame()
chunks = []
for year in range(from_year, to_year+1):

    rides_one_year = load_and_process_taxi_data(year)

    chunks.append(rides_one_year)
    break

# Concatenate all chunks at the end
rides = pd.concat(chunks, ignore_index=True)
print("Data loading complete.")

ts_data = transform_raw_data_into_ts_data(rides)

Logging into Hopsworks, and specifically into project CDA500P1

api_key = os.getenv('HOPSWORKS_API_KEY')  
project_name = os.getenv('HOPSWORKS_PROJECT_NAME')  

# pip install confluent-kafka
# Initialize connection to Hopsworks  
project = hopsworks.login(  
    api_key_value=api_key,  
    project=project_name  
)  
print(f"Successfully connected to Hopsworks project: {project_name}")

Setting up pointers, Feature groups

feature_store = project.get_feature_store()
FEATURE_GROUP_NAME = "time_series_hourly_feature_group"
FEATURE_GROUP_VERSION = 1

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time series data at hourly freaquency",
    primary_key=["pickup_location_id", "pickup hour"],
    event_time="pickup_hour"
)
feature_group.insert(ts_data, write_options={"wait_for_job": False})

df_memory_mb = rides.memory_usage(deep=True).sum() / (1024 * 1024)  
print(f"DataFrame size: {df_memory_mb:.2f} MB")
ts_data.info()
