==> notebook/18_plot_mae.ipynb <==
%load_ext autoreload
%autoreload 2
import sys
import os
import hopsworks
import pandas as pd
from datetime import timedelta
import plotly.express as px


sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

import src.config as config
from src.inference import get_feature_store, fetch_predictions


def fetch_hourly_rides(hours):
    current_hour = (pd.Timestamp.now(tz="Etc/UTC") - timedelta(hours=hours)).floor('h')

    fs = get_feature_store()
    fg = fs.get_feature_group(
        name=config.FEATURE_GROUP_NAME,
        version=1
    )

    query = fg.select_all()
    query = query.filter(fg.pickup_hour >= current_hour)

    return query.read()

df = fetch_hourly_rides(12)

df.sort_values(['pickup_hour'])

df_pred = fetch_predictions(12)

df_pred

merged_df = pd.merge(df, df_pred, on=['pickup_location_id', 'pickup_hour'])

merged_df

merged_df['difference'] = merged_df['predicted_demand'] - merged_df['rides']

merged_df.sort_values(["pickup_location_id", "pickup_hour"])

merged_df

df1 = df
df2 = df_pred

# Merge the DataFrames on 'pickup_location_id' and 'pickup_hour'  
merged_df = pd.merge(df1, df2, on=['pickup_location_id', 'pickup_hour'])  

# Calculate the absolute error  
merged_df['absolute_error'] = abs(merged_df['predicted_demand'] - merged_df['rides'])  

# Group by 'pickup_hour' and calculate the mean absolute error (MAE)  
mae_by_hour = merged_df.groupby('pickup_hour')['absolute_error'].mean().reset_index()  
mae_by_hour.rename(columns={'absolute_error': 'MAE'}, inplace=True)  

# Create a Plotly plot  
fig = px.line(  
    mae_by_hour,  
    x='pickup_hour',  
    y='MAE',  
    title='Mean Absolute Error (MAE) by Pickup Hour',  
    labels={'pickup_hour': 'Pickup Hour', 'MAE': 'Mean Absolute Error'},  
    markers=True  
)  

# Show the plot  
fig.show()

mae_by_hour["MAE"].mean()
